Methods

Index

PhysicalMeshes.center

PhysicalMeshes.centroid

PhysicalMeshes.circumcenter

PhysicalMeshes.incircle

PhysicalMeshes.insphere

PhysicalMeshes.len

PhysicalMeshes.midpoint

PhysicalMeshes.orient

PhysicalMeshes.orientation

PhysicalParticles.area

PhysicalParticles.volume

Basic geometry

PhysicalMeshes.len — Function

len(line::AbstractLine)

Length of line

len(t::AbstractTriangle)

Circumference of the triangle

PhysicalParticles.area — Function

area(c::AbstractCube2D)

Signed area of 2D cube

area(t::AbstractTriangle2D)
area(t::AbstractTriangle3D)

Unsigned area of triangle

PhysicalParticles.volume — Function

volume(c::AbstractCube3D)

Signed volume of 3D cube

volume(t::AbstractTetrahedron)

Unsigned volume of tetrahedron

PhysicalMeshes.center — Function

center(line::AbstractLine) = centroid(line)

Average position of line vertices.

PhysicalMeshes.centroid — Function

centroid(line::AbstractLine)

Average position of line vertices.

centroid(t::AbstractTriangle)

Averaged position center of triangle

centroid(t::AbstractTetrahedron)

Averaged position center of the tetrahedron

PhysicalMeshes.midpoint — Function

midpoint(line::AbstractLine) = centroid(line)

Average position of line vertices.

PhysicalMeshes.orient — Function

function orient(a::AbstractPoint2D, b::AbstractPoint2D, c::AbstractPoint2D)

Computes the orient.
The result is also a rough approximation of twice the signed area.

function orient(a::AbstractPoint3D, b::AbstractPoint3D, c::AbstractPoint3D)

Computes the orient.
The result is also a rough approximation of twice the signed area.

function orient(a::AbstractPoint3D, b::AbstractPoint3D, c::AbstractPoint3D, d::AbstractPoint3D)

Computes the orient.
The result is also a rough approximation of six times the signed volume.

orient(t::AbstractTriangle)

Orientation vector of triangle (return a PVector). See also orientation

orient(t::AbstractTetrahedron)

Return a positive number if positively oriented. Return a negative number if negatively oriented. Return zero if not oriented (indicating the tetrahedron has zero volume)

PhysicalMeshes.orientation — Function

orientation(t::AbstractTriangle2D)

Orientation of triangle. Return a trait (NegativelyOriented, PositivelyOriented, or UnOriented)

orientation(t::AbstractTetrahedron)

Orientation of tetrahedron. Return a trait (NegativelyOriented, PositivelyOriented, or UnOriented)

PhysicalMeshes.circumcenter — Function

circumcenter(t::AbstractTriangle)

Return the circumcenter of triangle (as a PVector or PVector2D)

circumcenter(t::AbstractTetrahedron)

Return the circumsphere center of tetrahedron

Predicates

PhysicalMeshes.incircle — Function

incircle(t::AbstractTriangle2D, p::AbstractPoint2D)

Test whether a point locates inside the circumcircle of triangle. Return a trait (Interior, Exterior, or OnEdge)

PhysicalMeshes.insphere — Function

insphere(t::AbstractTetrahedron, p::AbstractPoint3D)

Test whether a point locates inside the circumsphere center of tetrahedron. Return a trait (Interior, Exterior, or OnEdge)

Mesh

Missing docstring.

Missing docstring for particle2mesh. Check Documenter's build log for details.

Missing docstring.

Missing docstring for assignmesh. Check Documenter's build log for details.

Missing docstring.

Missing docstring for mesh2particle. Check Documenter's build log for details.

Missing docstring.

Missing docstring for assignparticle. Check Documenter's build log for details.